Revert "Move other dadguide enums into common (#1152)"
This reverts commit d97cf0699191dba09ca07c365913fdefaaae5820.
