+++ dadguide/common/enums.py
@@ -6,56 +6,6 @@ SERVERS = ["NA", "COMBINED"]
 Server = Literal["COMBINED", "NA"]
 
 
-class Attribute(Enum):
-    """Standard 5 PAD colors in enum form. Values correspond to Gungho values."""
-    Fire = 0
-    Water = 1
-    Wood = 2
-    Light = 3
-    Dark = 4
-    Unknown = 5
-    Nil = 6
-
-
-class MonsterType(Enum):
-    Evolve = 0
-    Balanced = 1
-    Physical = 2
-    Healer = 3
-    Dragon = 4
-    God = 5
-    Attacker = 6
-    Devil = 7
-    Machine = 8
-    Awoken = 12
-    Enhance = 14
-    Vendor = 15
-
-
-class InternalEvoType(Enum):
-    """The official GH evo types."""
-    Base = "Base"
-    Normal = "Normal"
-    Ultimate = "Ultimate"
-    Reincarnated = "Reincarnated"
-    Assist = "Assist"
-    Pixel = "Pixel"
-    SuperReincarnated = "Super Reincarnated"
-
-
-class Server(Enum):
-    JP = 0
-    NA = 1
-    KR = 2
-
-
-class AwakeningRestrictedLatent(Enum):
-    """Latent awakenings with availability gated by having an awakening"""
-    UnmatchableClear = 606
-    SpinnerClear = 607
-    AbsorbPierce = 608
-
-
 class AwokenSkills(Enum):
     ENHANCEDHP = 1
     ENHANCEDATK = 2
@@ -135,10 +85,3 @@ class AwokenSkills(Enum):
     LIGHTCOMBOCOUNT = 76
     DARKCOMBOCOUNT = 77
     CROSSATTACK = 78
-
-
-def enum_or_none(enum, value, default=None):
-    if value is not None:
-        return enum(value)
-    else:
-        return default
