+++ dadguide/models/monster_model.py
@@ -1,5 +1,8 @@
 from .base_model import BaseModel
-from ..common.enums import Attribute, MonsterType, AwakeningRestrictedLatent, enum_or_none
+from .enum_types import Attribute
+from .enum_types import MonsterType
+from .enum_types import AwakeningRestrictedLatent
+from .enum_types import enum_or_none
 from .active_skill_model import ActiveSkillModel
 from .leader_skill_model import LeaderSkillModel
 import tsutils
