+++ dadguide/token_mappings.py
@@ -2,7 +2,8 @@ from collections import defaultdict, namedtuple
 from enum import Enum
 from typing import Mapping, Tuple, TypeVar
 
-from .common.enums import AwokenSkills, Attribute, MonsterType
+from .common.enums import AwokenSkills
+from .models.enum_types import Attribute, MonsterType
 
 K = TypeVar('K')
 V = TypeVar('V')
